+++ src/main/java/com/keepgoing/keepserver/domain/user/payload/request/UserProfileDto.java
@@ -2,7 +2,7 @@ package com.keepgoing.keepserver.domain.user.payload.request;
 
 import com.keepgoing.keepserver.domain.book.payload.response.BookResponseDto;
 import com.keepgoing.keepserver.domain.device.payload.response.DeviceResponseDto;
-import com.keepgoing.keepserver.domain.user.entity.user.User;
+import com.keepgoing.keepserver.domain.user.domain.entity.user.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
