+++ src/main/java/com/keepgoing/keepserver/domain/user/security/service/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepserver.domain.user.security.service;
 
-import com.keepgoing.keepserver.domain.user.entity.user.User;
-import com.keepgoing.keepserver.domain.user.repository.user.UserRepository;
+import com.keepgoing.keepserver.domain.user.domain.entity.user.User;
+import com.keepgoing.keepserver.domain.user.domain.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
